Alter the C API, sketch out the python API
I'm now creating the results matrix from the python code, specifying
sites instead of windows (we'll layer these back in when we implement LD
scores and branch stats).

I've also implemented a python API, however for some reason the results
get added to the array every time we run the function, I'm not sure if
it's memory re-use and I need to zero things out or if it's a reference
issue and I am re-using a reference. (I suspect the former).