Checkpoint some ideas
Right now, the shape of the output data is (row, col, ss) and we want it
to be (ss, row, col). Also, we're still computing all samples under
every mutation, then selecting the subset of rows and columns from that.
We want to compute for only the subset.

There are some ideas worth checkpointing, but I want to simplify this
before proceeding.

Also, selecting rows and columns does not work.