Implement a method for specifying row/col sites
This allows a user to specify an array of row sites to be compared
against an array of column sites. To avoid computing redundant entries
in the matrix, we implement an index that tracks items and their
overlap. When we encounter entries that are redundant, we reflect their
value across the diagonal of the matrix (for the valid square sub-matrix
that crosses the diagonal). I've attached a diagram to the pull request
that describes this in more detail.

In addition to these changes, I've cleaned up the results matrix
allocation. Since the python code will be allocating and managing this
memory, we'll pull out the allocation code.

I've added tests to capture all of this new functionality.