Add python multipopulation tests
Fix state/result dimensions to properly output the number of dimensions
required for the number of sample sets in a statistic that is computed
between multiple sample sets.

Add some basic tests to ensure that between population tests are working,
generally. These tests only work for stats that return the same result
as their single-population counterparts.