bin boundaries; return bincount; optional ratemap
This change fixes the bin boundaries. Now, all bin boundaries for sites
will be left closed right open. The boundaries for branches will be
closed on both sides, minimizing the number of spanning bins. This is to
avoid integrating over a region of 0 length, which would affect the
bincount over a region that integrates to 0. Since we're performing an
integration, it doesn't matter if the left/right boundaries are opened
or closed.

Finally, add support for specifying a recombination map. I've performed
some minimal testing, but I'm not quite sure if it's working correctly.
I have a suspicion that there's a bug somewhere. It will allow the user
to specify bin boundaries in cM. Requires the tree to be simulated with
the same recombination map.